% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_mesh_ahull_crop}
\alias{build_mesh_ahull_crop}
\title{Create a grid within the alpha hull of a set of points.}
\usage{
build_mesh_ahull_crop(coords, npts, mesh.opts = list(alpha = 0.3),
  verbose = FALSE, ...)
}
\arguments{
\item{coords}{A matrix or data.frame of coordinates with two columns}

\item{npts}{The number of points before cropping to the alpha hull shape}

\item{mesh.opts}{A list with component \code{alpha} that controls the
shape of the alpha hull}

\item{...}{other arguments are silently ignored}
}
\value{
The coordinates of a grid of points as a \code{data.frame} with
        approximately \code{npts} rows and \code{ncol(coords)} columns.
}
\description{

}
\seealso{
Other mesh building functions: \code{\link{build_mesh_ahull_fill}};
  \code{\link{build_mesh_grid_proportional}}
}

