% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_mesh_grid_proportional}
\alias{build_mesh_grid_proportional}
\title{Create a grid with a number of points on each dimension
       proportional to the length of that dimension.}
\usage{
build_mesh_grid_proportional(coords, npts, pad = 0, ...)
}
\arguments{
\item{coords}{A matrix or data.frame of coordinates with two columns}

\item{npts}{The approximate total number of points of the output grid}

\item{pad}{Padding on each dimension (a positive number makes a grid
that is larger than the ranges of the coordinates).}

\item{...}{other arguments are silently ignored}
}
\value{
The coordinates of a grid of points as a \code{data.frame} with
        approximately \code{npts} rows and \code{ncol(coords)} columns.
}
\description{

}
\seealso{
Other mesh building functions: \code{\link{build_mesh_ahull_crop}};
  \code{\link{build_mesh_ahull_fill}}
}

