% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meadow.R
\docType{data}
\name{meadow}
\alias{meadow}
\title{Partial Data from a vegetation survey in a Californian subalpine meadow.}
\format{A data frame with 217 rows and 4 variables:
  \describe{
    \item{lon}{approximate longitude of the plot}
    \item{lat}{approximate latitude of the plot}
    \item{meanvwc}{the mean volumetric water content of the superficial soil}
    \item{tcover}{the total cover of vegetation in the plot}
  }}
\source{
\url{http://www.diamondse.info/}
}
\usage{
meadow
}
\description{
Part of the results of a vegetation survey carried out in Yosemite, CA.
}
\references{
A Multi-scale Approach to Searching for Disturbance Effects: Pack 
  Stock and Sierra Nevada Meadows. Steven R. Lee, Eric L. Berlow, Steven M. 
  Ostoja, John R. Matchett, Alexandre Genin, Stephen C. Hart, Matthew L. 
  Brooks. Plos One (2017)
}
\keyword{datasets}
