% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_grid_ahull_fill.R
\name{build_grid_ahull_fill}
\alias{build_grid_ahull_fill}
\title{Create a grid within the alpha hull of a set of points.}
\usage{
build_grid_ahull_fill(coords, npts, grid.opts = list(alpha = 0.3, error.tol =
  0.05, run.max = 20), verbose = FALSE, ...)
}
\arguments{
\item{coords}{A matrix or data.frame of coordinates with two columns}

\item{npts}{The approximate number of points of the requested grid}

\item{grid.opts}{A list with component \code{alpha} that controls the
shape of the alpha hull, \code{error.tol} the error
tolerance (in percentage) on the number of points and
\code{run.max} the maximum number of iterations.}

\item{verbose}{Show plots of the grid being built}

\item{...}{other arguments are silently ignored}
}
\value{
The coordinates of a grid of points as a \code{data.frame} with
        approximately \code{npts} rows and \code{ncol(coords)} columns.
}
\description{

}
\seealso{
Other grid building functions: \code{\link{build_grid_ahull_crop}};
  \code{\link{build_grid_proportional}}
}

