% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rollply}
\alias{rollply}
\title{Applies a function over a rolling window.}
\usage{
rollply(.data, .rollvars, wdw.size, fun, npts = 200, mesh = NULL,
  padding = "none", .parallel = FALSE, ...)
}
\description{
Applies a function over a rolling window.
}

