% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollply.R
\name{rollply}
\alias{rollply}
\title{Rollply}
\usage{
rollply(.data, .rollvars, wdw.size, fun, grid = NULL, grid.npts = 200,
  grid.type = "identical", grid.options = NULL, padding = "none",
  .parallel = FALSE, ...)
}
\arguments{
\item{.data}{data.frame to be processed}

\item{.rollvars}{variables describing the moving window: a formula of the
form ~ a + b | c, where a and b denote the variables used
for the rolling window and c an optional grouping variable}

\item{wdw.size}{window size}

\item{fun}{function to apply to each subset}

\item{grid}{data.frame of points at which the computation is done. If
\code{NULL} then a grid is generated using \code{grid.npts}
and \code{grid.type}.}

\item{grid.npts}{if grid is unspecified, then the number of points for the
resolution of the grid to build. Otherwise ignored.}

\item{grid.type}{The type of grid to generate}

\item{grid.options}{Options to be passed to the grid-generating function}

\item{padding}{padding policy outside of data range, one of 'none',
'outside', 'inside', or a numeric value}

\item{.parallel}{whether to use parallel processing (see \link{ddply} for
more information on parallelism).}

\item{...}{other arguments passed to ddply and fun}
}
\value{
A named data.frame with the function results at each grid point.
}
\description{
For each subset of a data.frame falling in a moving window, computes the
results of a function on this subset, then combine results in a data.frame.
}
\details{
Rollply allows applying a function on subsets falling in a window moving over
one or more variables. It is built internally over ddply and follows a very
similar syntax.
}
\examples{
library(plyr)
# Generate 1D random walk
dat <- data.frame(time=seq.int(1000),
                  position=cumsum(rnorm(1000,0,10)))

rollav <- rollply(dat, ~ time, wdw.size=10,
                  summarise, position=mean(position))

plot(position ~ time, data=dat)
lines(rollav, col='red')


# see http://github.com/alexgenin/rollply for more examples
}

