% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rollply}
\alias{rollply}
\title{Rollply}
\usage{
rollply(.data, .rollvars, wdw.size, fun, mesh = NULL, mesh.res = 200,
  padding = "none", .parallel = FALSE, ...)
}
\arguments{
\item{.data}{data.frame to be processed}

\item{.rollvars}{variables describing the moving window: a formula of the
form ~ a + b | c, where a and b denote the variables used
for the rolling window and c an eventual grouping}

\item{wdw.size}{window size}

\item{fun}{function to apply to each piece}

\item{mesh}{mesh of points at which the moving window is evaluated}

\item{mesh.res}{if mesh is unspecified, then the number of points for the
resolution of the default mesh to build. Otherwise ignored.}

\item{padding}{padding policy outside of data range, one of 'none',
'outside', 'inside', or a numeric value}

\item{.parallel}{whether to use parallel processing (see \link{ddply} for
more information on parallelism)}

\item{...}{other arguments passed first to ddply and/or adply, then to fun}
}
\description{
For each subset of a data.frame falling in a moving window, computes the
results of a function on this subset, then combine results in a data.frame.
}

